+++ changelog.md
@@ -1,3 +1,14 @@
+# v3.3.0
+
+## Changes 
+
+### Patch 9.2
+
+Updated for patch 9.2.0.
+
+Re-added Armor Token button/response group option.
+Added item data for Sepulcher of the First Ones.
+
 ## Bugfixes
 
 - *Fixed issue with `/rc remove` command.*